+++ CHANGELOG.md
@@ -9,7 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 - Added CODEOWNERS file;
-- Added `events/{event}/{command}/{resource}/internal` config key.
+- Added `events/{event}/{command}/{resource}/internal` config key;
+- Added `onbeforesave` event.
 
 ### Fixed